parametrized-locale-aware routing
For translation-aware routers, i.e. TranslationAwareTreeRouteStack:

If parameter „locale“ is detected during routing (either matching or
assembling), the translator will use this locale instead of a default
one. By doing so, the following route is possible:

   /:locale/{NEWS}/{TODAY}

Which would match/assemble with/to:

   /en/news/today
   /de/neues/heute

with translations (excerpt for locale ‚de‘: ‚NEWS‘ => ‚neues‘).