Ensure MiddlewareListener tests against abstract factories
zend-servicemanager v3 modified the behavior of `has()` to **not**
search abstract factories by default. You can force it to do so by
passing an optional second argument, a boolean flag, with a value of
boolean true.

The `DispatchListener` was already doing this, but the
`MiddlewareListener` was not. This patch adds tests to do so, and
updates some mocking assumptions to expect the flag.

Several method names in the `DispatchListener` tests were also updated to
clarify the behavior being tested.