Ensure that helpers with override factories work for all aliases
This patch implements tests for each of the url, basepath, and doctype
helpers, ensuring that the factory is invoked and the helper returned
properly configured *for each alias* we define, as well as for the fully
qualifed class name.

Interestingly, while the basepath and doctype tests fail for every case,
the URL helper only fails for the case of a fully qualified name; I'm
hoping this will still work as a fix.

I've refactored the `ViewHelperManagerFactory` substantially using the
extract method refactor to extract:

- helper factories (these are returned by dedicated methods now)
- helper factory registration
- helper configuration from other components

In the case of the second bullet point, I am now registering the plugins
using the FQCN as well as the normalized version of it, and not the
aliases; the aliases will resolve to the factory name.