Duck-type ServiceLocatorAwareInterface
Since zend-servicemanager v3 does not define
`ServiceLocatorAwareInterface`, then we cannot extend it and still support
that release series. As such, `AbstractController` no longer explicitly
implements it, but does so implicitly by defining the methods.

The initializers in both `ServiceManagerConfig` and `ControllerManager`
were updated to *also* look for the combination:

- interface `ServiceLocatorAwareInterface` does not exist AND
- method `setServiceLocator()` is present

If that combination is present, the same behavior is retained, including
the deprecation notices.