Ensure listener detachment works cross-version
Check if a returned callback is a CallbackHandler, and, if so, pull the actual
PHP callable from it before testing for the class. To ensure both BC and
forwards compatibility, ensure that the priority is stored when detaching, so
that re-attaching will continue to work.

In order to use `composer update --prefer-lowest`, I had to:

- update PHPUnit to `^4.5`, as that's the first version using prophecy, which we're
  using in tests now.
- add sebastian/version:^1.0.4, as that's the first version that fixes version
  autodiscovery, ensuring that the test runner knows we're on a version later
  than 2.5.