Updated to use zend-router v3
This patch updates the component to use zend-router 3.0, allowing the
removal of the entire `Zend\Mvc\Router` namespace, as well as related
factories in the `Zend\Mvc\Service` namespace.

Tests were updated to reflect the change. One set of tests, for the url
controller plugin, were marked skipped, as they rely on zend-view being
updated to use zend-router.

TODO
----

- [ ] once zend-view is updated, update the dependencies in this patch.
- [ ] experiment with having tests use zend-router as a module/config
  provider, to see if they can pass.