Updated test bootstrap
Current test failures are due to sebastian/version not reporting the correct
version when the lowest stable version is installed (1.0.4 of that package
corrected the issue). Since `test/bootstrap.php` was essentially unrequired at
this point, removing it in favor of using only the composer autoloader
simplifies dependencies while retaining the testing strategy.