Update zend-mvc v2 to use zend-servicemanager-di
zend-servicemanager-di contains tests and code to ensure it works against both
v2 and v3 versions of zend-servicemanager, unlike the code currently in
zend-mvc. Updating to use it is backwards compatible, assuming the following:

- zend-servicemanager-di is tested against both v2 and v3 of
  zend-servicemanager, demonstrating that the various classes it contains that
  overlap with zend-servicemanager v2 pose no conflicts.
- each of DiAbstractServiceFactoryFactory, DiFactory,
  DiServiceInitializerFactory, and DiStrictAbstractServiceFactory from zend-mvc
  v2 return artifacts from the `Zend\ServiceManager\Di` namespace, allowing them
  to simply extend the relevant factories that exist in zend-servicemanager-di.
- DiStrictAbstractServiceFactoryFactory can be left intact, ensuring it returns
  the `Zend\Mvc\Service\DiStrictAbstractServiceFactory` instance, and thus
  preserving the inheritance tree.

All of the DI-related factories in zend-mvc are now marked deprecated as well,
with annotations detailing how to upgrade code to work with
zend-servicemanager-di to make it forwards-compatible.