New LazyControllerFactory
Inspired by http://circlical.com/blog/2016/3/9/preparing-for-zend-f,
this abstract factory/factory can be used to create controller
instances for controllers defining constructor dependencies, using
the following rules:

- A parameter named `$config` typehinted as an array will receive the
  application "config" service (i.e., the merged configuration).
- Parameters type-hinted against array, but not named `$config` will
  be injected with an empty array.
- Scalar parameters will be resolved as null values.
- If a service cannot be found for a given typehint, the factory will
  raise an exception detailing this.
- Some services provided by Zend Framework components do not have
  entries based on their class name (for historical reasons); the
  factory contains a map of these class/interface names to the
  corresponding service name to allow them to resolve.

`$options` passed to the factory are ignored in all cases, as we cannot
make assumptions about which argument(s) they might replace.

As it implements zend-servicemanager's v3 AbstractFactoryInterface, it
may be used as either an abstract factory, or by mapping controller
class names to the factory.