Use a factory to create the SendResponseListener
Reported in:

- zendframework/zend-mvc-console#10
- zendframework/zend-mvc-console#11
- zendframework/zend-mvc-console#12

The `SendResponseListener` was lazy-instantiating an event manager on
first request to `getEventManager()`. However, because initializers run
after delegators, this meant that the EM instance composed did not have
a shared EM instance, which triggered the initializer to re-inject,
losing any listeners injected by delegators.

This patch introduces a factory for the `SendResponseListener`. The
factory creates the instance, and then injects it with an EM instance
pulled from the container; since these are guaranteed to have a shared
EM instance, the initializer will skip injection, keeping any listeners
injected by delegators.