Ensure that we replace `zendframework/zend-router` due to class conflicts
`zendframework/zend-mvc` in version less than 3.0 is still replacing `zendframework/zend-router` due to incompatible exception changes. Therefore we should tell projects that this package is replacing `zendframework/zend-router`.
The downsides are the mentioned exception incompatibilities but even if a user is requiring `zendframework/zend-router` in version 2 and using this package as well, he wont get the expected exceptions since `zendframework/zend-mvc` is being used instead of `zendframework/zend-router`.