Vary PluginManager implementation based on zend-servicemanager version detected
`PluginManager` overrides the `get()` method, which has a different
signature in v2 than in v3. Until PHP 7.2, this was not a problem;
however, in 7.2, even optional values must follow the exact same
signature as the parent. As a result, we now need to vary the
implementations.

This patch provides `AbstractPluginManager`, which contains the bulk of the
logic for implementing the `PluginManager`. It then introduces the
following:

- `PluginManagerSM2`, which extends `AbstractPluginManager` and
  implements the zend-servicemanager v2 `get()` signature.
- `PluginManagerSM3`, which extends `AbstractPluginManager` and
  implements the zend-servicemanager v3 `get()` signature.

It removes the `PluginManager` class.

It then adds a file-based autoloader that checks to see if
`Zend\ServiceManager\PluginManagerInterface` exists. If it does, it
aliases `PluginManager` to `PluginManagerSM3`; otherwise, it aliases it
to `PluginManagerSM2`.

Additionally, this patch updates to zend-servicemanager 2.7.10, which
has important fixes to how it handles creation options with
InvokableFactory that are necessary here.