Optimize Code
When calling `create()` outside of the `processPostData()` method (analog to `processBodyContent`) we can override the method to change the behaviour of parsing POST-Params and also getting the result back and not the response of the action. All other cases `GET`, `PUT`, `PATCH` and `DELETE` also calls their similar method in directly in the `onDispatch()` method.

These changes doesn't infect the workflow. It's just a small code optimization.

Signed-off-by: Martin Seysen <ms@clickbits.de>