Update quick-start.md
Currently the snippet illustrating modifications to application/module.config.php is incorrect relative to the module.config.php that has been downloaded withthe skeletonapplication asinstructed at the top of the tutorial.  This leaves the novice tutorial user facing some not helpful php errors.    

The modification included here shows the exact additions to the module.config.php in the context of the template file that the user starts with.

Thanks.